moving stuff back into OnCreateManager, because it actually does get injected from the start
