+++ workers/unity/Assets/Playground/Scripts/Bootstrap.cs
@@ -90,7 +90,7 @@ namespace Playground
             }
             
             var worker = Worker.Connect(config, new ForwardingDispatcher(), origin);
-            Instantiate(Level, Vector3.zero, Quaternion.identity);
+            Instantiate(Level, origin, Quaternion.identity);
             foreach (var system in SystemConfig.GetSystems(config.WorkerType))
             {
                 worker.World.GetOrCreateManager(system);
