--- workers/unity/Packages/com.improbable.gdk.core/Worker/TranslationUnitRegistry.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 20aa50dee10d344dd8ebd45964f50b67
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
